app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
pydantic==2.10.4
itsdangerous==2.2.0
jinja2==3.1.4
python-multipart==0.0.20
paramiko==3.5.0
